requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.1
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-dialogflow==2.1.2
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.37.1
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.14
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
